further extract models into their own files
