+++ shared/src/main/scala/petclinic/models/Vet.scala
@@ -1,18 +1,6 @@
 package petclinic.models
 
 import zio.json._
-import zio._
-
-import java.util.UUID
-
-final case class VetId(id: UUID) extends AnyVal
-
-object VetId {
-  implicit val codec: JsonCodec[VetId] = JsonCodec[UUID].transform(VetId(_), _.id)
-
-  def fromString(id: String): Task[VetId] =
-    ZIO.attempt(VetId(UUID.fromString(id)))
-}
 
 final case class Vet(id: VetId, lastName: String, specialty: String)
